+++ post/remove.php
@@ -34,7 +34,7 @@ class PostRemove extends BusinessLayer
 													WHERE idPost = :idPost
 														AND user_idUser = :user_idUser");
 														
-				if($statement->execute($params))
+				if($statement->execute($params) && $statement->rowCount() == 1)
 				{  
 					$statement = $this->m_db->prepare("DELETE FROM post
