+++ tag/remove.php
@@ -36,7 +36,7 @@ class TagRemove extends BusinessLayer
 													WHERE idTag = :idTag
 														AND user_idUser = :user_idUser");
 														
-				if($statement->execute($params))
+				if($statement->execute($params) && $statement->rowCount() == 1)
 				{  
 					$statement = $this->m_db->prepare("DELETE FROM tag
